organize repo with ctk and marimo dependency choice
